Update README instructions
This commit addresses several minor issues that exist in the README:
- A different intro section was used that did not specify the need
for Maven or Gradle, since the wrappers are being used.
- The instructions on building a jar had the wrong artifact listed.
Since building a jar is not required here and the testing instructions
are a bit long, the jar section was removed.
- Removed instructions about SpringBootApplication annotation,
since this is not mentioned in the guide.
- Removed links at the bottom of the page that result in a 404.
- The project id was not set and the links were not rendering
correctly.